+++ src/stats.h
@@ -46,7 +46,6 @@ struct ChannelStats {
 struct StatsBaton {
   // Input
   sharp::InputDescriptor *input;
-  VipsAccess accessMethod;
 
   // Output
   std::vector<ChannelStats> channelStats;